Move active anchor stats maintenance into its own submodule
This PR just moves code. This makes it easier to separate the actual
activity counting from activity unrelated maintenance. This is done
in preparation for the domain specific anchor activity stats.
